Refactor drive sensor infrastructure and drift
Reorganise drive module layout: split sensor feedback into
sensors::data (data plane) and sensors::control (control plane),
replace feedback.rs and lifecycle.rs with the new modules, and add a
sensors mod. Extract pure drift calculations into drift::math (remove
compensation.rs) and rename/reshape drift module. Add handlers.rs for
DriveLoop command handling and update imports across the drive tree.
